+++ ConsoleApplication1/point.cpp
@@ -15,10 +15,10 @@ void printPoint(point_t& point) {
 	std::cout << "end_y: " << point.end_y << std::endl;
 }
 
-double pointDifference(point_t& firstPoint, point_t& secondPoint) {
-	double dist = sqrt(pow(firstPoint.x_coordinate - secondPoint.x_coordinate, 2.0) + pow(firstPoint.y_coordinate - secondPoint.y_coordinate, 2.0)) + 
-		sqrt(pow(firstPoint.speed_x - secondPoint.speed_x, 2.0) + pow(firstPoint.speed_y - secondPoint.speed_y, 2.0)) + 
-		sqrt(pow(firstPoint.start_x - secondPoint.start_x, 2.0) + pow(firstPoint.start_y - secondPoint.start_y, 2.0)) + 
-		sqrt(pow(firstPoint.end_x - secondPoint.end_x, 2.0) + pow(firstPoint.end_y - secondPoint.end_y, 2.0));
+float pointDifference(point_t& firstPoint, point_t& secondPoint) {
+	float dist = (float)sqrt(pow(firstPoint.x_coordinate - secondPoint.x_coordinate, 2.0) + pow(firstPoint.y_coordinate - secondPoint.y_coordinate, 2.0)) + 
+		(float)sqrt(pow(firstPoint.speed_x - secondPoint.speed_x, 2.0) + pow(firstPoint.speed_y - secondPoint.speed_y, 2.0)) + 
+		(float)sqrt(pow(firstPoint.start_x - secondPoint.start_x, 2.0) + pow(firstPoint.start_y - secondPoint.start_y, 2.0)) + 
+		(float)sqrt(pow(firstPoint.end_x - secondPoint.end_x, 2.0) + pow(firstPoint.end_y - secondPoint.end_y, 2.0));
 	return dist;
 }
